refactor: restructure imports and socket messages
