+++ app/GX_integration/workflows/workflow_invocation.py
@@ -16,7 +16,8 @@ path.append('.')
 from bioblend.galaxy.objects.wrappers import Workflow, Invocation, History, Dataset, DatasetCollection
 
 from app.galaxy import GalaxyClient
-from app.api.socket_manager import SocketManager, SocketMessageType, SocketMessageEvent
+from app.api.socket_manager import SocketManager
+from app.api.socket_enums import SocketMessageEvent, SocketMessageType
 from app.enumerations import NumericLimits, InvocationTracking, JobState
 
 class WorkflowInvocationHandler:
