+++ app/GX_integration/workflows/worklfow_installer.py
@@ -12,7 +12,8 @@ path.append('.')
 from bioblend.galaxy.toolshed import ToolShedClient
 
 from app.galaxy import GalaxyClient
-from app.api.socket_manager import SocketManager, SocketMessageType, SocketMessageEvent
+from app.api.socket_manager import SocketManager
+from app.api.socket_enums import SocketMessageEvent, SocketMessageType
 from app.enumerations import NumericLimits
 
 class WorkflowInstaller:
