+++ app/api/socket_manager.py
@@ -1,6 +1,5 @@
 import asyncio
 import logging
-from enum import Enum
 from typing import Dict, List
 from fastapi import WebSocket, WebSocketDisconnect
 
@@ -15,46 +14,6 @@ def singleton(cls):
         return instances[cls]
     
     return get_instance
-
-class SocketMessageType(str, Enum):
-    """
-    Enumeration of message types used for socket communication
-    in workflow upload and execution, and tool execution processes.
-    """
-
-    # workflow upload
-    TOOL_INSTALL = "TOOL_INSTALL"
-    UPLOAD_WORKFLOW = "UPLOAD_WORKFLOW"
-    UPLOAD_FAILURE = "UPLOAD_FAILURE"
-    UPLOAD_COMPLETE = "UPLOAD_COMPLETE"
-    
-    # Workflow Invocation
-    INVOCATION_FAILURE = "INVOCATION_FAILURE"
-    INVOCATION_STEP_UPDATE = "INVOCATION_STEP_UPDATE"
-    INVOCATION_COMPLETE = "INVOCATION_COMPLETE"
-
-    #workflow execution
-    WORKFLOW_EXECUTE = "WORKFLOW_EXECUTE"
-    WORKFLOW_FAILURE = "WORKFLOW_FAILURE"
-
-    # Tool Execution
-    JOB_UPDATE = "JOB_UPDATE"
-    JOB_COMPLETE = "JOB_COMPLETE"
-    JOB_FAILURE = "JOB_FAILURE"
-    TOOL_EXECUTE = "TOOL_EXECUTE"
-
-class SocketMessageEvent(str, Enum):
-    """Enumeration of possible socket message event types for workflow and tool execution."""
-
-    # ping
-    ping = "ping"
-    # Workflow Execution
-    workflow_execute = "workflow_execute"
-    # Workflow Execution
-    workflow_upload = "workflow_upload"
-    # Tool Execution
-    tool_execute = "tool_execute"
-
 
 @singleton
 class SocketManager:
